+++ src/libes/ES_Synow_Source.cc
@@ -33,11 +33,11 @@
 #include <iostream>
 #include <algorithm>
 
-ES::Synow::Source::Source( ES::Synow::Grid& grid, int const mu_size ) :
+ES::Synow::Source::Source( ES::Synow::Grid& grid, size_t const mu_size ) :
     ES::Synow::Operator( grid ),
     _mu_size( mu_size )
 {
-    int v_size  = _grid->v_size;
+    size_t v_size  = _grid->v_size;
     _mu    = new double [ v_size * _mu_size * 2 ];
     _dmu   = new double [ v_size * _mu_size * 2 ];
     _shift = new double [ v_size * _mu_size * 2 ];
@@ -58,18 +58,18 @@ void ES::Synow::Source::operator() ( const ES::Synow::Setup& setup )
     double v_phot  = setup.v_phot;
     double v_outer = setup.v_outer;
     double v_step  = _grid->v[ 1 ] - _grid->v[ 0 ];
-    int    v_size  = _grid->v_size;
-    int    wl_used = _grid->wl_used;
+    size_t v_size  = _grid->v_size;
+    size_t wl_used = _grid->wl_used;
 
     // Prepare the wavelength-independent source metadata.
 
-    for( int iv = 0; iv < v_size; ++ iv )
+    for( size_t iv = 0; iv < v_size; ++ iv )
     {
 
         double v = _grid->v[ iv ];
         double mu_crit = iv > 0 ? sqrt( 1.0 - v_phot * v_phot / v / v ) : 0.0;
 
-        int    offset, i;
+        size_t offset, i;
         double dmu, mu_init;
 
         // Photosphere.
@@ -78,7 +78,7 @@ void ES::Synow::Source::operator() ( const ES::Synow::Setup& setup )
         mu_init = 1.0 - 0.5 * dmu;
 
         offset = iv * _mu_size * 2;
-        for( int im = 0; im < _mu_size; ++ im )
+        for( size_t im = 0; im < _mu_size; ++ im )
         {
             i = offset + im;
             _mu   [ i ] = mu_init - im * dmu;
@@ -92,7 +92,7 @@ void ES::Synow::Source::operator() ( const ES::Synow::Setup& setup )
         mu_init = mu_crit - 0.5 * dmu;
 
         offset = iv * _mu_size * 2 + _mu_size;
-        for( int im = 0; im < _mu_size; ++ im )
+        for( size_t im = 0; im < _mu_size; ++ im )
         {
             i = offset + im;
             _mu   [ i ] = mu_init - im * dmu;
@@ -106,7 +106,7 @@ void ES::Synow::Source::operator() ( const ES::Synow::Setup& setup )
 
     {
         double blue_wl = _grid->wl[ 0 ];
-        for( int i = 0; i < v_size * _mu_size; ++ i )
+        for( size_t i = 0; i < v_size * _mu_size; ++ i )
         {
             double shift_wl = _grid->wl[ 0 ] * _shift[ i ];
             if( shift_wl < blue_wl ) blue_wl = shift_wl;
@@ -116,13 +116,13 @@ void ES::Synow::Source::operator() ( const ES::Synow::Setup& setup )
 
     // Integration.
 
-    int      offset, im, i, start, ib, il, iu;
+    size_t   offset, im, i, start, ib, il, iu;
     double   v, in, d, vd, cl, cu, et, ss;
 
-    for( int iw = 0; iw < wl_used; ++ iw )
+    for( size_t iw = 0; iw < wl_used; ++ iw )
     {
         #pragma omp parallel for private( v, offset, im, i, in, start, ib, d, vd, il, iu, cl, cu, et, ss ) schedule( static, 1 )
-        for( int iv = 0; iv < v_size; ++ iv )
+        for( size_t iv = 0; iv < v_size; ++ iv )
         {
             v = _grid->v[ iv ];
             offset = iv * _mu_size * 2;
@@ -135,7 +135,7 @@ void ES::Synow::Source::operator() ( const ES::Synow::Setup& setup )
                 {
                     d  = ( _grid->wl[ iw ] / _grid->wl[ ib ] - 1.0 ) * 299.792;
                     vd = sqrt( v * v + d * d - 2.0 * v * d * _mu[ i ] );
-                    il = int( ( vd - v_phot ) / v_step );
+                    il = size_t( ( vd - v_phot ) / v_step );
                     iu = il + 1;
                     cl = ( _grid->v[ iu ] - vd ) / v_step;
                     cu = 1.0 - cl;
